Refactor doc comment synthetic service layout
