--- src/core/src/comments/doc-comment/service/synthetic.ts
@@ -1,3 +1,0 @@
-export * from "./synthetic/helpers.js";
-export * from "./synthetic/generation.js";
-export * from "./synthetic/merge.js";
