+++ src/core/test/comments/doc-comment/legacy-returns.test.ts
