+++ src/core/test/comments/doc-comment/manager.test.ts
