+++ src/core/test/comments/doc-comment/type-normalization.test.ts
